dumb changes

+++ ValueTables.py
@@ -13,7 +13,7 @@ class Heuristics:
         [10, 10, 15, 15, 15, 15, 10, 10],
         [60, 50, 50, 50, 50, 50, 50, 60],
         [0, 0, 0, 0, 0, 0, 0, 0]
-    ])
+    ], dtype=np.float)
 
     BLACK_PAWN_TABLE = np.array([
 
@@ -26,7 +26,7 @@ class Heuristics:
         [10, 10, 5, -5, -5, 5, 10, 10],
         [0, 0, 0, 0, 0, 0, 0, 0]
 
-    ])
+    ], dtype=np.float)
 
     KNIGHT_TABLE = np.array([
 
@@ -39,7 +39,7 @@ class Heuristics:
         [-5, 5, 5, 5, 5, 5, 5, -5],
         [-5, -5, -5, -5, -5, -5, -5, -5]
 
-    ])
+    ], dtype=np.float)
 
     BISHOP_TABLE = np.array([
 
@@ -51,7 +51,7 @@ class Heuristics:
         [0, 5, 10, 10, 10, 10, 5, 0],
         [0, 5, 5, 5, 5, 5, 5, 0],
         [0, 0, 0, 0, 0, 0, 0, 0]
-    ])
+    ], dtype=np.float)
 
     ROOK_TABLE = np.array([
         [10, 10, 10, 10, 10, 10, 10, 10],
@@ -62,7 +62,7 @@ class Heuristics:
         [10, 5, 0, 0, 0, 0, 5, 10],
         [10, 5, 5, 5, 5, 5, 5, 10],
         [10, 10, 10, 10, 10, 10, 10, 10]
-    ])
+    ], dtype=np.float)
 
     QUEEN_TABLE = np.array([
         [0, 0, 0, 0, 0, 0, 0, 0],
@@ -73,7 +73,7 @@ class Heuristics:
         [0, 5, 10, 10, 10, 10, 5, 0],
         [0, 5, 5, 5, 5, 5, 5, 0],
         [0, 0, 0, 0, 0, 0, 0, 0]
-    ])
+    ], dtype=np.float)
 
     KING_TABLE = np.array([
         [10, 15, 5, 5, 5, 5, 15, 10],
@@ -84,4 +84,4 @@ class Heuristics:
         [5, 0, 0, 0, 0, 0, 0, 5],
         [10, 10, 10, 5, 5, 10, 10, 10],
         [10, 15, 5, 5, 5, 5, 15, 10],
-    ])
+    ], dtype=np.float)
